final code cleanup

+++ src/ui/admin_view.py
@@ -1,10 +1,12 @@
+'''this module is responsible for admin view'''
 from tkinter import ttk, constants
-from services.recycle_service import recycle_service
 from tkinter.messagebox import askyesno
+from services.recycle_service import recycle_service
 
 
 class AdminView:
-    '''this class is responsible for the start view of the app'''
+    '''this class is responsible for the admin view of the app'''
+
     def __init__(self, root, handle_start):
         self._root = root
         self._password_entry = None
@@ -22,12 +24,8 @@ class AdminView:
         '''destroy the frame'''
         self._frame.destroy()
 
-    def _admin_handler(self):
-        password = self._password_entry.get()
-        recycle_service.login_admin(password)
-        self._handle_list()
-
     def show_users(self):
+        '''show all registered users'''
         self._show_users_button.destroy()
         self._logout_button.destroy()
         users_label = ttk.Label(master=self._frame, text="Users:")
@@ -38,30 +36,36 @@ class AdminView:
             user_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
         self._logout_button = ttk.Button(
             master=self._frame, text="Logout", command=self._handle_start)
-        self._logout_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
-        
+        self._logout_button.grid(
+            columnspan=2, sticky=constants.EW, padx=5, pady=5)
+
     def show_stats(self):
+        '''shows amount of recycled materials of all users'''
         self._show_stats_button.destroy()
         self._logout_button.destroy()
-        recycle_amount_label = ttk.Label(master=self._frame, text="Amount recycled by all users")
+        recycle_amount_label = ttk.Label(
+            master=self._frame, text="Amount recycled by all users")
         recycle_amount_label.grid(padx=5, pady=5, column=2)
-        list_of_recyclables = ["Bottles and cans","Cardboard","Electronics","Glass","Metal","Plastic","Paper","Batteries","Clothes"]
-        for i in range(0,8):
-            recyclable_label = ttk.Label(master=self._frame, text=(list_of_recyclables[i], "recycled:"))
-            recyclable_amount_label = ttk.Label(master=self._frame, text=recycle_service.recycle_list_all()[i])
+        list_of_recyclables = ["Bottles and cans", "Cardboard", "Electronics",
+                               "Glass", "Metal", "Plastic", "Paper", "Batteries", "Clothes"]
+        for i in range(0, 8):
+            recyclable_label = ttk.Label(master=self._frame, text=(
+                list_of_recyclables[i], "recycled:"))
+            recyclable_amount_label = ttk.Label(
+                master=self._frame, text=recycle_service.recycle_list_all()[i])
             recyclable_label.grid(padx=5, pady=5, column=1)
             recyclable_amount_label.grid(padx=5, pady=5, column=2)
         self._logout_button = ttk.Button(
             master=self._frame, text="Logout", command=self._handle_start)
-        self._logout_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
-
-    def del_user(self, username):
-        recycle_service.remove_user(username)
+        self._logout_button.grid(
+            columnspan=2, sticky=constants.EW, padx=5, pady=5)
 
     def confirm(self, username):
-        tempname = askyesno(title="confirmation", message='Are you sure you want to delete this user?')
-        if tempname:
-            self.del_user(username)
+        '''confirm that admin wants to delete a user'''
+        deleteconfirmation = askyesno(
+            title="confirmation", message='Are you sure you want to delete this user?')
+        if deleteconfirmation:
+            recycle_service.remove_user(username)
 
     def _initialize(self):
         self._frame = ttk.Frame(master=self._root)
@@ -70,7 +74,7 @@ class AdminView:
 
         self._show_users_button = ttk.Button(
             master=self._frame, text="Show users", command=self.show_users)
-        
+
         self._show_stats_button = ttk.Button(
             master=self._frame, text="Show statistics", command=self.show_stats)
 
@@ -78,8 +82,11 @@ class AdminView:
             master=self._frame, text="Logout", command=self._handle_start)
 
         heading_label.grid(columnspan=2, sticky=constants.W, padx=5, pady=5)
-        self._show_users_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
-        self._show_stats_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
-        self._logout_button.grid(columnspan=2, sticky=constants.EW, padx=5, pady=5)
+        self._show_users_button.grid(
+            columnspan=2, sticky=constants.EW, padx=5, pady=5)
+        self._show_stats_button.grid(
+            columnspan=2, sticky=constants.EW, padx=5, pady=5)
+        self._logout_button.grid(
+            columnspan=2, sticky=constants.EW, padx=5, pady=5)
 
         self._frame.grid_columnconfigure(1, weight=1, minsize=200)
